+++ p2p/index.js
@@ -1,0 +1,4 @@
+const makePeerRequest = require('./make_peer_request');
+const servicePeerRequests = require('./service_peer_requests');
+
+module.exports = {makePeerRequest, servicePeerRequests};